+++ packages/shared/lib/core/layer-2/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './addGasBudget'
 export * from './isLayer1Destination'
-export * from './validateEthereumAddress'
