+++ packages/shared/lib/core/network/utils/isShimmerNetworkProtocol.ts
@@ -1,4 +1,4 @@
-import { NetworkProtocol } from '@core/network'
+import { NetworkProtocol } from '@core/network/enums'
 
 export function isShimmerNetworkProtocol(networkProtocol: NetworkProtocol): boolean {
     return networkProtocol === NetworkProtocol.Shimmer
