+++ packages/shared/lib/core/nfts/stores/all-account-nfts.store.ts
@@ -1,4 +1,5 @@
 import { writable } from 'svelte/store'
+
 import { INft } from '../interfaces'
 
 export const allAccountNfts = writable<INft[][]>([])
