+++ packages/shared/lib/core/utils/crypto/classes/index.ts
@@ -1,0 +1,2 @@
+export * from './bech32.class'
+export * from './bech32-helper.class'
