+++ packages/shared/lib/core/utils/crypto/constants/ethereum-address-hash-size.constant.ts
@@ -1,0 +1,1 @@
+export const ETHEREUM_ADDRESS_HASH_SIZE = 256
