+++ packages/shared/lib/core/utils/crypto/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './ethereum-address-hash-size.constant'
