+++ packages/shared/lib/core/utils/crypto/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './validateBech32Address'
+export * from './validateEthereumAddress'
