+++ packages/shared/lib/core/utils/crypto/utils/validateEthereumAddress.ts
@@ -1,11 +1,12 @@
-import { Keccak } from 'sha3'
 import { get } from 'svelte/store'
 
-import { localize } from '@core/i18n'
-import { networkHrp } from '@core/network'
-import { validateBech32Address } from '@core/utils'
+import { Keccak } from 'sha3'
 
-const HASH_SIZE = 256
+import { localize } from '@core/i18n'
+import { networkHrp } from '@core/network/stores'
+
+import { ETHEREUM_ADDRESS_HASH_SIZE } from '../constants'
+import { validateBech32Address } from './validateBech32Address'
 
 export function validateEthereumAddress(address: string): string {
     // TODO: currently, validateBech32Address throws an error string if an invalid address is provided.
@@ -31,7 +32,7 @@ export function validateEthereumAddress(address: string): string {
 // Check if EIP-55 mixed-case checksum address encoding is followed (https://eips.ethereum.org/EIPS/eip-55)
 function isChecksumAddress(address: string): string {
     address = address.replace('0x', '')
-    const keccak256 = new Keccak(HASH_SIZE)
+    const keccak256 = new Keccak(ETHEREUM_ADDRESS_HASH_SIZE)
     const addressHash = keccak256.update(address.toLowerCase()).digest('hex')
     for (let i = 0; i < address.length; i++) {
         // The nth letter should be uppercase if the nth digit of casemap is 1
