+++ packages/shared/lib/core/wallet/actions/activities/preprocessOutputsForAccount.ts
@@ -1,8 +1,10 @@
-import { IAccountState } from '@core/account'
-import { preprocessGroupedOutputs } from '@core/wallet/utils/outputs/preprocessGroupedOutputs'
 import { OutputData, Transaction } from '@iota/wallet'
-import { IProcessedTransaction } from '../../interfaces'
 import { IOutputResponse, ITransactionPayload } from '@iota/types'
+
+import { IAccountState } from '@core/account/interfaces'
+
+import { preprocessGroupedOutputs } from '../../utils/outputs'
+import { IProcessedTransaction } from '../../interfaces'
 
 export async function preprocessOutputsForAccount(account: IAccountState): Promise<IProcessedTransaction[]> {
     const outputs = await account.outputs()
