+++ packages/shared/lib/core/wallet/utils/convertHexAddressToBech32.ts
@@ -1,7 +1,8 @@
-import { networkHrp } from '@core/network'
-import { Bech32Helper } from '@core/utils'
-import { Converter } from '@core/utils'
 import { get } from 'svelte/store'
+
+import { networkHrp } from '@core/network/stores'
+import { Converter } from '@core/utils/convert'
+import { Bech32Helper } from '@core/utils/crypto'
 
 export function convertHexAddressToBech32(addressType: number, hexAddress: string): string {
     return hexAddress
