+++ packages/shared/lib/core/wallet/utils/getBech32AddressFromAddressTypes.ts
@@ -1,6 +1,7 @@
-import { convertHexAddressToBech32 } from '@core/wallet'
-import { ADDRESS_TYPE_ALIAS, ADDRESS_TYPE_ED25519, ADDRESS_TYPE_NFT } from '@core/wallet/constants'
 import type { AddressTypes } from '@iota/types'
+
+import { convertHexAddressToBech32 } from './convertHexAddressToBech32'
+import { ADDRESS_TYPE_ALIAS, ADDRESS_TYPE_ED25519, ADDRESS_TYPE_NFT } from '../constants'
 
 export function getBech32AddressFromAddressTypes(address: AddressTypes): string {
     switch (address?.type) {
